Added cast to avoid floating-point buffer count
git-svn-id: https://svn.blender.org/svnroot/verse/branches/release-r6-branch@3723 49c17534-1d02-dc11-b283-0030488c597c

+++ v_cmd_buf.c
@@ -34,7 +34,7 @@ void cmd_buf_init(void)
 	for(i = 0; i < VCMDBS_COUNT; i++)
 	{
 		VCMDBufData.buffers[i] = NULL;
-		VCMDBufData.available[i] = vcmdbuf_chunk_size[i] * VCMDBUF_INIT_CHUNK_FACTOR;
+		VCMDBufData.available[i] = (unsigned int) (vcmdbuf_chunk_size[i] * VCMDBUF_INIT_CHUNK_FACTOR);
 		for(j = 0, buf = NULL; j < VCMDBufData.available[i]; j++, buf = b)
 		{
 			b = v_cmd_buf_allocate(i);
